feat: obchod skutečná data
