+++ ui/src/main.ts
@@ -1,6 +1,10 @@
 import { initEnv } from "./env";
 import { renderPages } from "./pages"
 
+/**
+ * TODO: kompletní refactor
+ */
+
 console.log("Preact starting ...")
 
 initEnv();
