+++ ui/src/pages/obchod/app.tsx
@@ -1,182 +1,55 @@
-import { FunctionComponent, render } from "preact";
-import { useCallback, useEffect, useMemo, useRef, useState } from "preact/hooks";
-import { Overlay } from "../../components/Overlay";
-import { GAMECON_KONSTANTY } from "../../env";
-import { ObchodMřížka } from "./components/ObchodMřížka";
-import { ObchodShrnutí } from "./components/ObchodShrnutí";
-import "./app.less";
-import {
-  DefiniceObchod,
-  DefiniceObchodMřížka,
-  DefiniceObchodMřížkaBuňka,
-  ObjednávkaPředmět,
-  Předmět,
-} from "../../api/obchod/types";
+import { createContext, FunctionComponent } from "preact";
+import { useCallback, useEffect, useState } from "preact/hooks";
+import { fetchMřížky, fetchPředměty } from "../../api/obchod/endpoints";
+import { DefiniceObchod, DefiniceObchodMřížka, Předmět } from "../../api/obchod/types";
+import { Obchod } from "./components/Obchod";
+import "./App.less";
 
-type TObchodProps = {};
+type TAppProps = {
 
-const mockdefinices: DefiniceObchod = {
-  mřížky: [
-    {
-      id: 1,
-      buňky: [
-        { typ: "předmět", předmět: { id: 2, text: "triko", zbývá: 2 } },
-        ...Array(12)
-          .fill(0)
-          .map(
-            (x) =>
-              ({
-                typ: "stránka",
-                id: 2,
-                text: "stranka2",
-                barvaPozadí: "lightblue",
-              } as DefiniceObchodMřížkaBuňka)
-          ),
-      ],
-    },
-    {
-      id: 2,
-      buňky: Array(12)
-        .fill(0)
-        .map((x) => ({ typ: "zpět" } as DefiniceObchodMřížkaBuňka)),
-    },
-  ],
 };
 
-/** Ruční optimalizace, nedoporučuju používat pokud neznáš dobře react! */
-const useFixed = <T,>(value: T) => useRef(value).current;
+// TODO: hodně společné logika pro Obchod a Obchod nastavení, sjednotit
 
-const usePředmětyObjednávka = () => {
-  const [předmětyObjednávka, setPředmětyObjednávka] = useState<
-    ObjednávkaPředmět[]
-  >([]);
+export const PředmětyContext = createContext<Předmět[]>([]);
 
-  const předmětPřidej = useFixed((předmět: Předmět) => {
-    setPředmětyObjednávka((předměty) =>
-      předměty.some((x) => x.předmět.id === předmět.id)
-        ? předměty.map((x) =>
-            x.předmět.id === předmět.id ? { ...x, množství: x.množství + 1 } : x
-          )
-        : předměty.concat([{ množství: 1, předmět }])
-    );
-  });
-  const předmětOdeber = useFixed((předmět: Předmět) => {
-    setPředmětyObjednávka((předměty) =>
-      předměty
-        .map((x) =>
-          x.předmět.id === předmět.id ? { ...x, množství: x.množství - 1 } : x
-        )
-        .filter((x) => x.množství >= 1)
-    );
-  });
-  const předmětySmažVšechny = useFixed(() => {
-    setPředmětyObjednávka([]);
-  });
-
-  return {
-    předmětyObjednávka,
-    předmětPřidej,
-    předmětOdeber,
-    předmětySmažVšechny,
-  };
+const usePředměty = () => {
+  const [předměty, setPředměty] = useState<Předmět[] | null | undefined>();
+  useEffect(() => {
+    (async () => {
+      setPředměty(await fetchPředměty());
+    })();
+  }, []);
+  return předměty;
 };
 
-const useMřižka = () => {
-  const [výchoZíMřížka, setVýchoZíMřížka] = useState(1);
-  const [mřížkaId, _setMřížkaId] = useState(1);
-  const [mřížkaIdHist, setMřížkaIdHist] = useState<number[]>([1]);
-  const setId = (id: number) => {
-    setMřížkaIdHist((x) => [id, ...x]);
-    _setMřížkaId(id);
-  };
-  const setZpět = () => {
-    const last = mřížkaIdHist[0] ?? 0;
-    setMřížkaIdHist((x) => x.slice(1));
-    setId(last);
-  };
 
-  const mřížka: DefiniceObchodMřížka | undefined = mockdefinices.mřížky.find(
-    (x) => x.id === mřížkaId
-  );
-
-  const setShrnutí = useFixed(() => setId(-1));
-
-  const setVýchozí = useCallback(() => setId(výchoZíMřížka), []);
-
-  const setMřížka = useMemo(() =>({
-    id: setId, zpět: setZpět, výchozí: setVýchozí, shrnutí: setShrnutí 
-  }), [setId, setZpět, setVýchozí, setShrnutí]);
-
-  return { mřížka, setMřížka };
-};
-
-export const Obchod: FunctionComponent<TObchodProps> = (props) => {
+export const App: FunctionComponent<TAppProps> = (props) => {
   const {} = props;
 
-  const [visible, setVisible] = useState(GAMECON_KONSTANTY.IS_DEV_SERVER);
+  const [definiceObchod, setDefiniceObchod] = useState<
+    DefiniceObchod | null | undefined
+  >();
 
-  const {
-    předmětyObjednávka: předměty,
-    předmětPřidej,
-    předmětOdeber,
-    předmětySmažVšechny,
-  } = usePředmětyObjednávka();
-  const { mřížka, setMřížka } = useMřižka();
+  const předměty = usePředměty();
 
   useEffect(() => {
-    window.preactMost.obchod.show = () => {
-      setVisible(true);
-    };
+    (async () => {
+      setDefiniceObchod(await fetchMřížky());
+    })();
   }, []);
 
-  const onBuňkaClicked = useCallback((buňka: DefiniceObchodMřížkaBuňka) => {
-    switch (buňka.typ) {
-      case "stránka":
-        setMřížka.id(buňka.id);
-        break;
-      case "zpět":
-        setMřížka.zpět();
-        break;
-      case "předmět":
-        setMřížka.id(0);
-        předmětPřidej(buňka.předmět);
-        break;
-    }
-  }, []);
-
-  const onDalšíPředmět = useCallback(() => {
-    setMřížka.výchozí();
-  }, []);
-
-  const onStorno = useCallback(() => {
-    předmětySmažVšechny();
-    setMřížka.výchozí();
-    setVisible(false);
-  }, []);
-
-  return (
+  return definiceObchod === null || předměty === null ? (
+    <div> nepodařilo se načíst nastavení mřížek !!! </div>
+  ) : definiceObchod === undefined || předměty === undefined ? (
+    <div>načítání ...</div>
+  ) : (
     <>
-      {visible ? (
-        <Overlay>
-          <div class="shop--container">
-            {mřížka ? (
-              <ObchodMřížka {...{ mřížka, onBuňkaClicked }} />
-            ) : (
-              <ObchodShrnutí
-                {...{
-                  předmětyObjednávka: předměty,
-                  onDalšíPředmět,
-                  onStorno,
-                  předmětPřidej,
-                  předmětOdeber,
-                }}
-              />
-            )}
-          </div>
-        </Overlay>
-      ) : undefined}
+      <PředmětyContext.Provider value={předměty}>
+        <Obchod definice={definiceObchod} />
+      </PředmětyContext.Provider>
     </>
   );
 };
 
-Obchod.displayName = "Obchod";
+App.displayName = "App";
