+++ ui/src/pages/obchod/components/Obchod/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Obchod";
