+++ ui/src/pages/obchod/components/ObchodMřížka/ObchodMřížka.tsx
@@ -1,5 +1,11 @@
 import { FunctionComponent } from "preact";
-import { DefiniceObchodMřížka, DefiniceObchodMřížkaBuňka } from "../../../../api/obchod/types";
+import { useContext } from "preact/hooks";
+import {
+  DefiniceObchodMřížka,
+  DefiniceObchodMřížkaBuňka,
+  DefiniceObchodMřížkaBuňkaPředmět,
+} from "../../../../api/obchod/types";
+import { PředmětyContext } from "../../App";
 import "./ObchodMřížka.less";
 
 type TObchodMřížkaProps = {
@@ -10,6 +16,8 @@ type TObchodMřížkaProps = {
 export const ObchodMřížka: FunctionComponent<TObchodMřížkaProps> = (props) => {
   const { onBuňkaClicked, mřížka: mřížka } = props;
 
+  const všechnyPředměty = useContext(PředmětyContext);
+
   return (
     <>
       <div class="shop-grid--container">
@@ -18,9 +26,17 @@ export const ObchodMřížka: FunctionComponent<TObchodMřížkaProps> = (props)
             <div
               onClick={() => onBuňkaClicked?.(x)}
               class={`shop-grid--item shop-grid--item-${i}`}
-              style={x.barvaPozadí ? {backgroundColor: x.barvaPozadí } : ""}
+              style={x.barvaPozadí ? { backgroundColor: x.barvaPozadí } : ""}
             >
-              <div>{!x.text && x.typ === "předmět" ? x.předmět.text : x.text}</div>
+              <div>
+                {!x.text && x.typ === "předmět"
+                  ? (
+                      všechnyPředměty.find(
+                        (y) => y.id === x.cilId
+                      )
+                    )?.název
+                  : x.text}
+              </div>
             </div>
           );
         })}
